Apply Clang-Tidy suggestions

+++ src/cybel/types/color.cpp
@@ -59,18 +59,18 @@ Color4f& Color4f::set_bytes(std::uint8_t rgb,std::uint8_t a) { return set_bytes(
 
 Color4f& Color4f::set_bytes(std::uint8_t r,std::uint8_t g,std::uint8_t b) {
   return set(
-    r / 255.0f,
-    g / 255.0f,
-    b / 255.0f
+    static_cast<float>(r) / 255.0f,
+    static_cast<float>(g) / 255.0f,
+    static_cast<float>(b) / 255.0f
   );
 }
 
 Color4f& Color4f::set_bytes(std::uint8_t r,std::uint8_t g,std::uint8_t b,std::uint8_t a) {
   return set(
-    r / 255.0f,
-    g / 255.0f,
-    b / 255.0f,
-    a / 255.0f
+    static_cast<float>(r) / 255.0f,
+    static_cast<float>(g) / 255.0f,
+    static_cast<float>(b) / 255.0f,
+    static_cast<float>(a) / 255.0f
   );
 }
 
@@ -82,12 +82,12 @@ Color4f Color4f::with_b(float b) const { return Color4f{r,g,b,a}; }
 
 Color4f Color4f::with_a(float a) const { return Color4f{r,g,b,a}; }
 
-Color4f Color4f::with_byte_r(std::uint8_t r) const { return Color4f{r / 255.0f,g,b,a}; }
+Color4f Color4f::with_byte_r(std::uint8_t r) const { return Color4f{static_cast<float>(r) / 255.0f,g,b,a}; }
 
-Color4f Color4f::with_byte_g(std::uint8_t g) const { return Color4f{r,g / 255.0f,b,a}; }
+Color4f Color4f::with_byte_g(std::uint8_t g) const { return Color4f{r,static_cast<float>(g) / 255.0f,b,a}; }
 
-Color4f Color4f::with_byte_b(std::uint8_t b) const { return Color4f{r,g,b / 255.0f,a}; }
+Color4f Color4f::with_byte_b(std::uint8_t b) const { return Color4f{r,g,static_cast<float>(b) / 255.0f,a}; }
 
-Color4f Color4f::with_byte_a(std::uint8_t a) const { return Color4f{r,g,b,a / 255.0f}; }
+Color4f Color4f::with_byte_a(std::uint8_t a) const { return Color4f{r,g,b,static_cast<float>(a) / 255.0f}; }
 
 } // Namespace.
